fix: collect username with emojis
