+++ packages/maskbook/src/utils/index.ts
@@ -30,3 +30,4 @@ export * from './theme'
 export * from './type'
 export * from './utils'
 export * from './watcher'
+export * from './collectNodeText'
